refactor: Refactor UseCase classes to use invoke operator
The `CopyToClipboardUseCase`, `InstallScannerUseCase`, and `CodeScanUseCase` classes have been refactored to use the `invoke` operator.
The lambda arguments of `install`, `scan` and `copy` methods has been moved to `invoke` operator arguments of `InstallScannerUseCase`, `CodeScanUseCase` and `CopyToClipboardUseCase` respectively.
The `onCopyToClipboardButtonClicked` has been changed to use the invoke operator of `copyToClipboardUseCase`.
The `onScanButtonClicked` has been changed to use the invoke operator of `codeScanUseCase`.
The `InstallScannerUseCase` and `CodeScanUseCase` has been changed to `installScannerUseCase` and `codeScanUseCase` respectively.
This change makes the code cleaner and more idiomatic.